+++ src/main/java/com/gotcha/_global/config/S3Config.java
@@ -10,7 +10,7 @@ import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
 
 @Configuration
-@Profile({"dev", "prod"})
+@Profile({"local", "dev", "prod"})
 public class S3Config {
 
     @Value("${aws.credentials.access-key}")